rename pieChart.ts -> d3PieChart.ts to avoid namespace issues
